errores en las querys de JPQL

+++ src/main/java/pe/edu/upc/wayrugrupo5/DTOS/IncidenciaPorCategoriaDTO.java
@@ -3,11 +3,11 @@ package pe.edu.upc.wayrugrupo5.DTOS;
 public class IncidenciaPorCategoriaDTO {
 
     private String nombreCategoria;
-    private long total;
+    private int total;
 
     public IncidenciaPorCategoriaDTO() {}
 
-    public IncidenciaPorCategoriaDTO(String nombreCategoria, long total) {
+    public IncidenciaPorCategoriaDTO(String nombreCategoria, int total) {
         this.nombreCategoria = nombreCategoria;
         this.total = total;
     }
@@ -19,10 +19,10 @@ public class IncidenciaPorCategoriaDTO {
         this.nombreCategoria = nombreCategoria;
     }
 
-    public long getTotal() {
+    public int getTotal() {
         return total;
     }
-    public void setTotal(long total) {
+    public void setTotal(int total) {
         this.total = total;
     }
 }
